+++ aggregate-snapshot/aggregate-snapshot-service/src/test/java/uk/gov/justice/services/core/aggregate/SnapshotAwareAggregateServiceIT.java
@@ -69,6 +69,7 @@ import uk.gov.justice.services.messaging.jms.DefaultEnvelopeConverter;
 import uk.gov.justice.services.messaging.jms.JmsEnvelopeSender;
 import uk.gov.justice.services.test.utils.persistence.DatabaseCleaner;
 import uk.gov.justice.subscription.ParserProducer;
+import uk.gov.justice.subscription.SubscriptionHelper;
 import uk.gov.justice.subscription.YamlFileFinder;
 import uk.gov.justice.subscription.domain.eventsource.EventSourceDefinition;
 import uk.gov.justice.subscription.registry.DefaultEventSourceDefinitionFactory;
@@ -204,7 +205,10 @@ public class SnapshotAwareAggregateServiceIT {
             QualifierAnnotationExtractor.class,
             SnapshotAwareEventSourceFactory.class,
 
-            DefaultEventSourceDefinitionFactory.class
+            DefaultEventSourceDefinitionFactory.class,
+
+            SubscriptionHelper.class
+
     })
 
     public WebApp war() {